requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.12
commonmark==0.9.1
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
multidict==6.0.2
Pygments==2.11.2
python-dotenv==0.20.0
requests==2.27.1
rfc3986==1.5.0
rich==12.0.1
sniffio==1.2.0
typing_extensions==4.1.1
urllib3==1.26.9
websocket-client==1.3.1
yarl==1.7.2
